Broadcast updates bwn tabs,windows and workers
