+++ src/classes/dexie/dexie.ts
@@ -42,7 +42,7 @@ import { Middleware, DexieStacks } from '../../public/types/middleware';
 import { virtualIndexMiddleware } from '../../dbcore/virtual-index-middleware';
 import { hooksMiddleware } from '../../hooks/hooks-middleware';
 import { IndexableType } from '../../public';
-import { observabilityMiddleware } from '../live-query/live-query';
+import { observabilityMiddleware } from '../live-query/observability-middleware';
 
 export interface DbReadyState {
   dbOpenError: any;
@@ -206,7 +206,10 @@ export class Dexie implements IDexie {
     // Default middlewares:
     this.use(virtualIndexMiddleware);
     this.use(hooksMiddleware);
-    this.use(observabilityMiddleware); 
+    this.use(observabilityMiddleware);
+
+    // Call all enablers
+    (Dexie as any as DexieConstructor).enablers.forEach(enabler => enabler(this));
 
     // Call each addon:
     addons.forEach(addon => addon(this));
